requirements.txt
discord.py
openpyxl
python-dotenv 
Flask
pandas
flask-cors 